feat(clean):update const
