+++ internal/datanode.go
@@ -15,6 +15,7 @@ import (
 	"sync"
 	"time"
 	"tinydfs-base/common"
+	"tinydfs-base/protocol/pb"
 	"tinydfs-base/util"
 )
 
@@ -295,6 +296,30 @@ type ChunkSendInfo struct {
 	SendType   int    `json:"send_type"`
 }
 
+func ConvChunkInfo(chunkInfos []*pb.ChunkInfo) []ChunkSendInfo {
+	chunkSendInfos := make([]ChunkSendInfo, len(chunkInfos))
+	for i := 0; i < len(chunkInfos); i++ {
+		chunkSendInfos[i] = ChunkSendInfo{
+			ChunkId:    chunkInfos[i].ChunkId,
+			DataNodeId: chunkInfos[i].DataNodeId,
+			SendType:   int(chunkInfos[i].SendType),
+		}
+	}
+	return chunkSendInfos
+}
+
+func DeConvChunkInfo(chunkSendInfos []ChunkSendInfo) []*pb.ChunkInfo {
+	chunkInfos := make([]*pb.ChunkInfo, len(chunkSendInfos))
+	for i := 0; i < len(chunkInfos); i++ {
+		chunkInfos[i] = &pb.ChunkInfo{
+			ChunkId:    chunkSendInfos[i].ChunkId,
+			DataNodeId: chunkSendInfos[i].DataNodeId,
+			SendType:   int32(chunkSendInfos[i].SendType),
+		}
+	}
+	return chunkInfos
+}
+
 // DegradeDataNode degrade a DataNode based on given stage. If DataNode is dead,
 // it will remove DataNode from dataNodeMap and put all Chunk's id in Chunks and
 // FutureSendChunks of the DataNode to pendingChunkQueue so that system can make
@@ -480,4 +505,49 @@ func GetAvgChunkNum() int {
 	}
 	avgChunkNum := count / len(dataNodeMap)
 	return avgChunkNum
+}
+
+// DoExpand gets the chunk copied according to this new dataNode.
+func DoExpand(dataNode *DataNode) int {
+	var (
+		pendingCount  = GetAvgChunkNum()
+		selfChunks    = dataNode.Chunks
+		pendingChunks = set.NewSet()
+		pendingMap    = map[string][]string{}
+	)
+For:
+	for {
+		notFound := true
+		for _, node := range dataNodeMap {
+			if node.status == common.Alive {
+				for chunk := range node.Chunks.Iter() {
+					if !pendingChunks.Contains(chunk) && !selfChunks.Contains(chunk) {
+						notFound = false
+						pendingChunks.Add(chunk)
+						if pendingDataNodeChunks, ok := pendingMap[node.Id]; ok {
+							pendingDataNodeChunks = append(pendingDataNodeChunks, chunk.(string))
+						} else {
+							pendingMap[node.Id] = []string{chunk.(string)}
+						}
+						if pendingChunks.Cardinality() == pendingCount {
+							break For
+						}
+						break
+					}
+				}
+			}
+		}
+		if notFound {
+			break
+		}
+	}
+	expandOperation := &ExpandOperation{
+		Id:           util.GenerateUUIDString(),
+		SenderPlan:   pendingMap,
+		ReceiverPlan: dataNode.Id,
+		ChunkIds:     util.Interfaces2TypeArr[string](pendingChunks.ToSlice()),
+	}
+	data := getData4Apply(expandOperation, common.OperationExpand)
+	GlobalMasterHandler.Raft.Apply(data, 5*time.Second)
+	return pendingChunks.Cardinality()
 }
